chore: refactor to support systems
- Particles are now contained within PSystems

+++ src/include/point2f.h
@@ -2,15 +2,15 @@
 #define H_POINT2F
 
 typedef struct Point2f {
-    float x;
-    float y;
+    double x;
+    double y;
 } Point2f;
 
-void p2f_init(Point2f *ptr, float x, float y);
+void p2f_init(Point2f *ptr, double x, double y);
 
-Point2f p2f_mul(Point2f *ptr, float mul);
+Point2f p2f_mul(Point2f *ptr, double mul);
 
 void p2f_add_inplace(Point2f *original, Point2f *add);
-void p2f_mul_inplace(Point2f *original, float mul);
+void p2f_mul_inplace(Point2f *original, double mul);
 
 #endif
